refactor(viewport): extract duplicated default visual reset logic into BaseVolumeViewport
