+++ packages/core/src/RenderingEngine/VolumeViewport.ts
@@ -17,7 +17,7 @@ import type {
   ViewReferenceSpecifier,
 } from '../types';
 import type { ViewportInput } from '../types/IViewport';
-import { actorIsA, isImageActor } from '../utilities/actorCheck';
+import { actorIsA } from '../utilities/actorCheck';
 import getClosestImageId from '../utilities/getClosestImageId';
 import getSliceRange from '../utilities/getSliceRange';
 import getSpacingInNormalDirection from '../utilities/getSpacingInNormalDirection';
@@ -26,8 +26,6 @@ import triggerEvent from '../utilities/triggerEvent';
 
 import BaseVolumeViewport from './BaseVolumeViewport';
 import setDefaultVolumeVOI from './helpers/setDefaultVolumeVOI';
-import { setTransferFunctionNodes } from '../utilities/transferFunctionUtils';
-import type { ImageActor } from '../types/IActor';
 import getImageSliceDataForVolumeViewport from '../utilities/getImageSliceDataForVolumeViewport';
 import { transformCanvasToIJK } from '../utilities/transformCanvasToIJK';
 import { transformIJKToCanvas } from '../utilities/transformIJKToCanvas';
@@ -838,25 +836,7 @@ class VolumeViewport extends BaseVolumeViewport {
     }
     setDefaultVolumeVOI(volumeActor.actor as vtkVolume, imageVolume);
 
-    if (isImageActor(volumeActor)) {
-      // Check if there are default properties (e.g., a 3D preset or colormap)
-      // to restore. This ensures "Reset" goes back to the configured look
-      // rather than the raw initial grayscale state.
-      const properties = this.getDefaultProperties(volumeId);
-
-      if (properties?.preset || properties?.colormap) {
-        this.setProperties(properties, volumeId, true);
-      } else {
-        const transferFunction = (volumeActor.actor as ImageActor)
-          .getProperty()
-          .getRGBTransferFunction(0);
-
-        setTransferFunctionNodes(
-          transferFunction,
-          this.initialTransferFunctionNodes
-        );
-      }
-    }
+    this._restoreDefaultVisualProperties(volumeActor, volumeId);
 
     const eventDetails = {
       ...super.getVOIModifiedEventDetail(volumeId),
